[Library.API] - Modified Order computed columns to float and its sql query.
[UserData.API]
- Modified Enrollment/Material API query.

+++ src/Services/Library/Library.API/Infrastructure/Entities/Lesson.cs
@@ -9,7 +9,7 @@ public class Lesson
 	public string Description { get; set; }
 	public int OrderNumerator { get; set; }
 	public int OrderDenominator { get; set; }
-	public decimal Order { get; set; }
+	public float Order { get; set; }
 
 	// Relationship
 	public int CourseId { get; set; }
